kconfig: Add wrapping of long expressions at ORs
Some symbols in configuration may be selected by very-very long boolean
expressions. Reading through such long expressions, it is extremely hard
to identify what exactly selects the given option.

This commit adds a new line every time a boolean OR is encountered in
the "outer" expression. That is, it won't break apart separate
sub-expressions like OPTION_A && (OPTION_B || OPTION_C), but will add a
new line before OPTION_C in an expression like:

OPTION_A && OPTION_B || OPTION_C

This commit also adds indentation of the wrapped lines to the label
position. That is, the expression on the new line will continue starting
at the same character position as it started in the first line, right
after the label such as "  Selected by:  ".

This commit also improves some wording in the code, as well as remove
some redundant string arithmetics, and adds an obviously forgotten
gettext support to the "  Selects: " label.

Signed-off-by: Alexander Amelkin <alexander@amelkin.msk.ru>
Cc: Yann E. MORIN <yann.morin.1998@free.fr>
Cc: linux-kbuild@vger.kernel.org